requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
pillow
pytesseract
pandas
openpyxl
psycopg2-binary
